fix(qa): sanitize gateway logs in thrown errors
