+++ extensions/qa-lab/src/gateway-rpc-client.ts
@@ -1,4 +1,5 @@
 import { formatErrorMessage } from "openclaw/plugin-sdk/error-runtime";
+import { formatQaGatewayLogsForError } from "./gateway-log-redaction.js";
 import { callGatewayFromCli } from "./runtime-api.js";
 
 type QaGatewayRpcRequestOptions = {
@@ -13,7 +14,7 @@ export type QaGatewayRpcClient = {
 
 function formatQaGatewayRpcError(error: unknown, logs: () => string) {
   const details = formatErrorMessage(error);
-  return new Error(`${details}\nGateway logs:\n${logs()}`);
+  return new Error(`${details}${formatQaGatewayLogsForError(logs())}`);
 }
 
 function runQueuedQaGatewayRpc<T>(queue: Promise<void>, task: () => Promise<T>) {
